requirements.txt
ipdb==0.13.9
pytest==7.2.0
wheel
plotnine==0.10.1
numpy==1.23.5
scipy==1.9.3
scikit-learn==1.2.0
scikit-tensor-py3 @ https://github.com/jonjoncardoso/scikit-tensor-py3/archive/master.zip
networkx==2.8.8
termcolor==2.1.1

### UTILS
jupyterlab==3.5.1
jupyterlab_execute_time==2.3.0
ipywidgets==8.0.3
tqdm==4.64.1
wget==3.2

pdoc3==0.10.0 #dev
openpyxl
fuzzywuzzy
pyyaml
<<<<<<< HEAD
seasborn
=======
seaborn
cartopy
contextily
geopandas
osmnx
>>>>>>> hsafdari_dev
